+++ formio_sale/__manifest__.py
@@ -9,7 +9,7 @@
     'author': 'Nova Code',
     'website': 'https://www.novacode.nl',
     'live_test_url': 'https://demo15.novacode.nl',
-    'category': 'Sales',
+    'category': 'Forms/Forms',
     'depends': [
         'sale_management',
         'formio',