+++ formio/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 16.0.1.1
+
+Fix deprecation warning in (http) controller `send_fonts_file`, by
+using Werkzeug's implementation.
+
 ## 16.0.1.0
 
 Initial release.